Dummy mtg card info fetcher
I made a function that pull random 50 price and rarity and return a Dataframe to test scryfall api.

I then used Seaborn library and tried to see how it work with Tkinter. For this i plotted Prices in function of Rarity in a boxplot.

Despite the low number of data (50 or less), it should show visual difference in the price median and range