Redis + Redrawing attempt
Redis : my aim is to store price of card during time. I installed a redis server on my local computer then built a function to get all id and return the date of the day and the price of the day via scryfall.

Trying to find way to update canvas with new graphs based on selection. Not working properly yet. Will properly rework, so i get a unique frame.